rename: Shut up a some compiler warnings
It's handy to decide on one type and then use that one. Casting size_t
to int64_t back and forth gets tedious and is also error prone, since
int64_t is (always) 64 bits and signed, while size_t is (always, on
GNU systems) 32 bits and unsigned.

+++ rename.c
@@ -19,7 +19,7 @@ typedef struct renames {
 
 FILE *new_file;
 
-size_t rename_len;
+int64_t rename_len;
 char **from_name;
 char **to_name;
